requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.5.6
botocore==1.8.20
bz2file==0.98
certifi==2017.11.5
chardet==2.3.0
cycler==0.10.0
Cython==0.27.3
Django==1.11.3
django-bootstrap3==9.0.0
django-extensions==1.8.1
docopt==0.6.2
docutils==0.14
docx2txt==0.6
EbookLib==0.15
fake-useragent==0.1.10
gensim==3.2.0
idna==2.6
jmespath==0.9.3
lxml==4.1.1
matplotlib==2.0.2
mysqlclient==1.3.12
nltk==3.2.5
numpy==1.13.3
olefile==0.44
pdfminer.six==20170720
pdftotext==1.1.0
Pillow==4.3.0
pipreqs==0.4.9
pycryptodome==3.4.7
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
python-memcached==1.58
python-pptx==0.6.5
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==0.18.1
six==1.10.0
smart-open==1.5.5
SpeechRecognition==3.6.3
textblob==0.15.0
Unidecode==0.4.21
urllib3==1.22
xlrd==1.0.0
XlsxWriter==1.0.2
yarg==0.1.9
